+++ .codesandbox/ci.json
@@ -8,7 +8,7 @@
     "packages/material-ui-lab",
     "packages/material-ui-styles",
     "packages/material-ui-system",
-    "packages/material-ui-theming",
+    "packages/material-ui-private-theming",
     "packages/material-ui-types",
     "packages/material-ui-utils",
     "packages/material-ui-unstyled",
@@ -23,7 +23,7 @@
     "@material-ui/styled-engine": "packages/material-ui-styled-engine/build",
     "@material-ui/styled-engine-sc": "packages/material-ui-styled-engine-sc/build",
     "@material-ui/system": "packages/material-ui-system/build",
-    "@material-ui/theming": "packages/material-ui-theming/build",
+    "@material-ui/private-theming": "packages/material-ui-private-theming/build",
     "@material-ui/types": "packages/material-ui-types/build",
     "@material-ui/utils": "packages/material-ui-utils/build",
     "@material-ui/unstyled": "packages/material-ui-unstyled/build"
