+++ docs/babel.config.js
@@ -24,7 +24,7 @@ const alias = {
   '@material-ui/styled-engine': '../packages/material-ui-styled-engine/src',
   // '@material-ui/styled-engine': '../packages/material-ui-styled-engine-sc/src',
   '@material-ui/system': '../packages/material-ui-system/src',
-  '@material-ui/theming': '../packages/material-ui-theming/src',
+  '@material-ui/private-theming': '../packages/material-ui-private-theming/src',
   '@material-ui/utils': '../packages/material-ui-utils/src',
   '@material-ui/unstyled': '../packages/material-ui-unstyled/src',
   docs: './',
