+++ docs/next.config.js
@@ -119,7 +119,7 @@ module.exports = {
                           '../packages/material-ui-styled-engine-sc/src',
                         '@material-ui/styles': '../packages/material-ui-styles/src',
                         '@material-ui/system': '../packages/material-ui-system/src',
-                        '@material-ui/theming': '../packages/material-ui-theming/src',
+                        '@material-ui/private-theming': '../packages/material-ui-private-theming/src',
                         '@material-ui/utils': '../packages/material-ui-utils/src',
                         '@material-ui/unstyled': '../packages/material-ui-unstyled/src',
                       },
