+++ packages/material-ui-private-theming/README.md
@@ -1,3 +1,3 @@
-# @material-ui/theming
+# @material-ui/private-theming
 
 Theming utilities to be used in both `@material-ui/styles` & `@material-ui/core`.
