+++ packages/material-ui-private-theming/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@material-ui/theming",
+  "name": "@material-ui/private-theming",
   "version": "5.0.0-alpha.32",
   "private": false,
   "author": "Material-UI Team",
@@ -14,7 +14,7 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/mui-org/material-ui.git",
-    "directory": "packages/material-ui-theming"
+    "directory": "packages/material-ui-private-theming"
   },
   "license": "MIT",
   "bugs": {
@@ -34,7 +34,7 @@
     "build:copy-files": "node ../../scripts/copy-files.js",
     "prebuild": "rimraf build",
     "release": "yarn build && npm publish build --tag next",
-    "test": "cd ../../ && cross-env NODE_ENV=test mocha 'packages/material-ui-theming/**/*.test.{js,ts,tsx}'",
+    "test": "cd ../../ && cross-env NODE_ENV=test mocha 'packages/material-ui-private-theming/**/*.test.{js,ts,tsx}'",
     "typescript": "tslint -p tsconfig.json \"{src,test}/**/*.{spec,d}.{ts,tsx}\" && tsc -p tsconfig.json"
   },
   "peerDependencies": {
