+++ packages/material-ui-private-theming/src/getThemeProps/getThemeProps.d.ts
