+++ packages/material-ui-private-theming/src/getThemeProps/index.d.ts
