+++ packages/material-ui-private-theming/src/getThemeProps/index.js
