+++ packages/material-ui-private-theming/src/propsToClassKey/index.d.ts
