+++ packages/material-ui-styles/src/ThemeProvider/index.d.ts
@@ -1,2 +1,2 @@
-export { default } from '@material-ui/theming/ThemeProvider';
-export * from '@material-ui/theming/ThemeProvider';
+export { default } from '@material-ui/private-theming/ThemeProvider';
+export * from '@material-ui/private-theming/ThemeProvider';
