+++ packages/material-ui-styles/src/defaultTheme/index.d.ts
@@ -1,3 +1,3 @@
 /* eslint-disable import/prefer-default-export */
 
-export { DefaultTheme } from '@material-ui/theming';
+export { DefaultTheme } from '@material-ui/private-theming';
