+++ packages/material-ui-styles/src/getThemeProps/index.d.ts
@@ -1,2 +1,2 @@
-export { default } from '@material-ui/theming/getThemeProps';
-export * from '@material-ui/theming/getThemeProps';
+export { default } from '@material-ui/private-theming/getThemeProps';
+export * from '@material-ui/private-theming/getThemeProps';
