+++ packages/material-ui-styles/src/useTheme/index.d.ts
@@ -1,2 +1,2 @@
-export { default } from '@material-ui/theming/useTheme';
-export * from '@material-ui/theming/useTheme';
+export { default } from '@material-ui/private-theming/useTheme';
+export * from '@material-ui/private-theming/useTheme';
