+++ packages/material-ui-styles/src/useTheme/index.js
@@ -1,1 +1,1 @@
-export { default } from '@material-ui/theming/useTheme';
+export { default } from '@material-ui/private-theming/useTheme';
