+++ packages/material-ui/package.json
@@ -63,7 +63,7 @@
     "@material-ui/styled-engine": "5.0.0-alpha.26",
     "@material-ui/styles": "5.0.0-alpha.32",
     "@material-ui/system": "5.0.0-alpha.31",
-    "@material-ui/theming": "5.0.0-alpha.32",
+    "@material-ui/private-theming": "5.0.0-alpha.32",
     "@material-ui/types": "6.0.0",
     "@material-ui/unstyled": "5.0.0-alpha.32",
     "@material-ui/utils": "5.0.0-alpha.31",
