+++ packages/material-ui/scripts/rollup.config.js
@@ -20,11 +20,11 @@ const nestedFolder = {
       return resolved;
     }
 
-    if (importee.indexOf('@material-ui/theming/') === 0) {
+    if (importee.indexOf('@material-ui/private-theming/') === 0) {
       const folder = importee.split('/')[2];
       const resolved = path.resolve(
         __dirname,
-        `../../../packages/material-ui-theming/src/${folder}/index.js`,
+        `../../../packages/material-ui-private-theming/src/${folder}/index.js`,
       );
       return resolved;
     }
