+++ packages/material-ui/src/SwipeableDrawer/SwipeableDrawer.js
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
 import { elementTypeAcceptingRef } from '@material-ui/utils';
-import { getThemeProps } from '@material-ui/theming';
+import { getThemeProps } from '@material-ui/private-theming';
 import Drawer, { getAnchor, isHorizontal } from '../Drawer/Drawer';
 import ownerDocument from '../utils/ownerDocument';
 import ownerWindow from '../utils/ownerWindow';
