+++ packages/material-ui/src/styles/ThemeProvider.js
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
-import { ThemeProvider as MuiThemeProvider } from '@material-ui/theming';
+import { ThemeProvider as MuiThemeProvider } from '@material-ui/private-theming';
 import { exactProp } from '@material-ui/utils';
 import { ThemeContext as StyledEngineThemeContext } from '@material-ui/styled-engine';
 import useTheme from './useTheme';
