+++ packages/material-ui/src/styles/experimentalStyled.js
@@ -1,5 +1,5 @@
 import styled from '@material-ui/styled-engine';
-import { propsToClassKey } from '@material-ui/theming';
+import { propsToClassKey } from '@material-ui/private-theming';
 import { unstable_styleFunctionSx as styleFunctionSx } from '@material-ui/system';
 import defaultTheme from './defaultTheme';
