+++ packages/material-ui/src/withWidth/withWidth.js
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
 import { getDisplayName } from '@material-ui/utils';
-import { getThemeProps } from '@material-ui/theming';
+import { getThemeProps } from '@material-ui/private-theming';
 import hoistNonReactStatics from 'hoist-non-react-statics';
 import useTheme from '../styles/useTheme';
 import { breakpointKeys } from '../styles/createBreakpoints';
