+++ scripts/sizeSnapshot/webpack.config.js
@@ -68,8 +68,8 @@ async function getWebpackEntries() {
       path: 'packages/material-ui-styles/build/index.js',
     },
     {
-      name: '@material-ui/theming',
-      path: 'packages/material-ui-theming/build/index.js',
+      name: '@material-ui/private-theming',
+      path: 'packages/material-ui-private-theming/build/index.js',
     },
     {
       name: '@material-ui/system',
@@ -166,7 +166,7 @@ module.exports = async function webpackConfig(webpack, environment) {
           ),
           '@material-ui/styles': path.join(workspaceRoot, 'packages/material-ui-styles/build'),
           '@material-ui/system': path.join(workspaceRoot, 'packages/material-ui-system/build'),
-          '@material-ui/theming': path.join(workspaceRoot, 'packages/material-ui-theming/build'),
+          '@material-ui/private-theming': path.join(workspaceRoot, 'packages/material-ui-private-theming/build'),
           '@material-ui/utils': path.join(workspaceRoot, 'packages/material-ui-utils/build'),
           '@material-ui/unstyled': path.join(workspaceRoot, 'packages/material-ui-unstyled/build'),
         },
