+++ tsconfig.json
@@ -26,8 +26,8 @@
       "@material-ui/styles": ["./packages/material-ui-styles/src"],
       "@material-ui/styles/*": ["./packages/material-ui-styles/src/*"],
       "@material-ui/system": ["./packages/material-ui-system/src"],
-      "@material-ui/theming": ["./packages/material-ui-theming/src"],
-      "@material-ui/theming/*": ["./packages/material-ui-theming/src/*"],
+      "@material-ui/private-theming": ["./packages/material-ui-private-theming/src"],
+      "@material-ui/private-theming/*": ["./packages/material-ui-private-theming/src/*"],
       "@material-ui/types": ["./packages/material-ui-types"],
       "@material-ui/unstyled": ["./packages/material-ui-unstyled/src"],
       "@material-ui/unstyled/*": ["./packages/material-ui-unstyled/src/*"],
